requirements.txt
google-genai>=1.0.0
pyaudio>=0.2.14
pyautogui>=0.9.54
pystray>=0.19.0
Pillow>=10.0.0
keyboard>=0.13.5
numpy>=1.24.0
